refactor(evidence): split CNCF code into pkg/evidence/cncf subpackage
Reshape pkg/evidence/ so the CNCF AI Conformance code lives in its
own subpackage, making room for future evidence kinds (notably the
recipe-test attestation pipeline tracked under #754):

  pkg/evidence/*  ->  pkg/evidence/cncf/

The top-level pkg/evidence/ now contains only doc.go, which describes
the layout. The CNCF subpackage keeps its existing public surface
(NewCollector, IsValidFeature, ...); imports in pkg/cli/validate.go
are updated to match. No behavior changes.
