+++ pkg/evidence/cncf/scripts/collect-evidence.sh
