+++ pkg/evidence/cncf/scripts/manifests/hpa-gpu-test.yaml
