+++ pkg/evidence/cncf/scripts/manifests/trainer-pytorch-test.yaml
