+++ pkg/evidence/cncf/templates.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package evidence
+package cncf
 
 const indexTemplate = `# CNCF AI Conformance Evidence
